refactor: isolate coordinate transforms
